refactor code for swapping puzzle pieces
