+++ client/src/reducers/initialState.js
@@ -1,7 +1,11 @@
-import { range } from 'lodash';
+import { range, shuffle, without } from 'lodash';
+
+import { EMPTY_TILE } from '../constants';
+
+const numbers = range(0, 16);
 
 export default {
   puzzleState: {
-    puzzleNumbers: range(0, 16),
+    puzzleNumbers: shuffle(without(numbers, EMPTY_TILE)).concat(EMPTY_TILE),
   },
 };
